rewrite unit tests
